Fixed signed/unsigned mismatch that new gcc version catches
git-svn-id: https://yuma.svn.sourceforge.net/svnroot/yuma/trunk@3 1ffd1d58-2efc-45df-af60-190a33ec9863

+++ netconf/src/ncx/cap.c
@@ -484,7 +484,7 @@ status_t
     }
 
     /* set the length and malloc flag */
-    cap->cap_mod_len = p - cap->cap_mod;
+    cap->cap_mod_len = (uint32)(p - cap->cap_mod);
     cap->cap_mod_malloc = NULL;
 
     /* check if stopped on a separator char,
